model_project

+++ GardenGreens/src/main/java/com/masai/model/Orders.java
@@ -21,7 +21,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-
 @Entity
 public class Orders {
     @Id
@@ -31,8 +30,10 @@ public class Orders {
 	private String transactionMode;
 	private Integer planterQuantity;
 	private Integer plantQuantity;
-	private Double cost;
+	private Integer cost;
+	
 	private Integer customerId;
+	
 	@OneToMany(cascade = CascadeType.ALL)
 	@JoinColumn(name = "orderId")
 	private Set<Planter> planters = new HashSet<>();
